Remove all ASL code from this branch; upstream branch is now the source
The ASL code has for some time been imported only on to the upstream
branch. Removing the ASL code from this branch (leaving only the
import script and documentation relating to this repo, neither of
which is part of ASL itself) helps avoid confusion and removes the
need to do rebasing of the master branch.

COPYING has been left as-is, GPL v2. The original commit for this repo
(before ASL was brought in) used GPL v1, as did ASL at the time, but
ASL changed to GPL v2 around 2000-10-16 (asl-current-142-bld9, commit
136aa8e here). I am assuming that Kuba was ok with changing the
license for his import script and README to v2 as well.

I also considered creating a new branch starting at the first commit
and containing only the import code/doc commits, which would replace
the current `master`, become the default branch (so the README shows
on GitHub's top page for the repo), and probably keep the name
`master`. However, it seemed to me better and less confusing to keep
the current history rather than, for existing clones of the repo, a)
forcing a rebase of the `master` branch or b) leaving them on a branch
that no longer exists in the remote.