Adding IRPN command
Like IRP, it repeats the instructions for all parameters it receives.
Unlike IRP, its first parameter is a number of argument names,
followed by this number of argument names, followed by
a list of argument values which the given names will evaluate
to.
irpn with count 1 is identical to irp; irpn with count 0
is invalid.

Usage:
    irpn <count>,<count names>,parameters
        ; instructions
    endm

Fixes #5.