68k: Fix bogus 'short address not allowed' errors
This bug could be triggered with the following code:

	CPU 68000
	phase $200000
	move.w	(dummy).w,d0
dummy = $100

What happens is that, when an undefined symbol is encountered in the
first pass, it is temporarily assigned the value of the program
counter.

In the above example, the `dummy` variable is undefined at the time
the line `move.w (dummy).w,d0` is encountered, causing it to be
temporarily evaluated to $200000. Since $200000 is not within short
addressing range, the assembler throws an error, despite the actual
value of the `dummy` variable being $100.

To fix this, I've made it so that the short addressing range check is
not performed if the symbol is undefined.