Fix forward references causing errors in displacements.
My code contains this line:
move.w	.command_names-FlgID__First*2(pc,d4.w),d4

'FlgID__First' depends on a constant that is defined after it, making
it a forward reference which must be resolved on a later pass. This
was being resolved to some arbitrary value which is larger than 0x7F,
causing AS to attempt to assemble an instruction with a 16-bit
displacement, but then it throws an error because 16-bit
displacements are not supported by the 68000.

There is already a mechanism present to avoid this kind of problem,
but it is only applied to 'questionable' values and not
forward-referenced ones. This commit fixes that.