Make out-of-range PC-relative operands print the correct error.
Out-of-range values were making AS try to use an operand type that
the 68000 does not support, which it would then complain about the
68000 not supporting. By forcing AS to use the correct operand type
when targetting the 68000, the range error is caught and a more
appropriate error message is printed instead.

This more-appropriate error-handling logic was completely unused as a
result of this bug!